src/requirements.txt
pip==21.3.1
beautifulsoup4==4.11.1
fake-useragent==0.1.11
kafka-python==2.0.2
pymongo==4.2.0
selenium==4.4.3
requests==2.28.1